add base for wishlist tab
